Changed exhaust mechanic.
Exhaust struct tracks each players exhaust in 4 bits, for 4 players
total 16 is used.

+++ src/player/side.rs
@@ -58,7 +58,7 @@ impl Side{
         }
 
     }
-    pub fn index(&self) -> usize{
+    pub fn index(&self) -> u8{
         match self{
             North => 0,
             East => 1,
